+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## 1.3.0
+* Add `SubSecondPrecision` option to handle Fluentd v0.14 (and v1) sub-second EventTime (default: false)
+* Add `WriteTimeout` option
+* Fix API of `Post` to accept `msgp.Marshaler` objects for better performance
+
+## 1.2.1
+* Fix a bug not to reconnect to destination twice or more
+* Fix to connect on background goroutine in async mode
+
+## 1.2.0
+* Add `MarshalAsJSON` feature for `message` objects which can be marshaled as JSON
+* Fix a bug to panic for destination system outage
+
 ## 1.1.0
  * Add support for unix domain socket
  * Add asynchronous client creation